+++ CHANGELOG.md
@@ -22,7 +22,11 @@ with respect to its command line interface and HTTP interface
   Previously the build succeeded, and pushed a new docker image, but subsequent
   deploys did not use the new docker image, since sous uses the image digest, not
   the docker tag itself to identify images with SourceIDs.
-	
+* Client: the runmount build strategy is detected based on Dockerfile
+  environment variables. Specifically, to be considered runmount, the Dockerfile
+  or its parents must declare the SOUS_RUN_IMAGE_SPEC and BUILD_OUT environment variables.
+  All known existing runmount containers already do this.
+
 ## [0.5.120](//github.com/opentable/sous/compare/0.5.117...0.5.120)
 ### Fixed
 * Deployments to PROD clusters will use Jenkins agents with 'mesos-prod-sc'